+++ src/Ops.Controllers/Areas/ContentManagement/ViewModels/ApplicationPermissions/IndexViewModel.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+using Ocuda.Ops.Models.Definitions.Models;
+using Ocuda.Utility.Models;
+
+namespace Ocuda.Ops.Controllers.Areas.ContentManagement.ViewModels.ApplicationPermissions
+{
+    public class IndexViewModel
+    {
+        public ICollection<DataWithCount<ApplicationPermissionDefinition>> ApplicationPermissions { get; set; }
+    }
+}