+++ src/Ops.Data/OpsContext.cs
@@ -34,6 +34,8 @@ namespace Ocuda.Ops.Data
                 .HasKey(_ => new { _.UserId, _.UserMetadataTypeId });
             modelBuilder.Entity<SectionCategory>()
                 .HasKey(_ => new { _.SectionId, _.CategoryId });
+            modelBuilder.Entity<PermissionGroupApplication>()
+                .HasKey(_ => new { _.PermissionGroupId, _.ApplicationPermission });
             modelBuilder.Entity<PermissionGroupPageContent>()
                 .HasKey(_ => new { _.PermissionGroupId, _.PageHeaderId });
             modelBuilder.Entity<PostCategory>()
@@ -63,6 +65,7 @@ namespace Ocuda.Ops.Data
         public DbSet<Link> Links { get; set; }
         public DbSet<LinkLibrary> LinkLibraries { get; set; }
         public DbSet<PermissionGroup> PermissionGroups { get; set; }
+        public DbSet<PermissionGroupApplication> PermissionGroupApplication { get; set; }
         public DbSet<PermissionGroupPageContent> PermissionGroupPageContents { get; set; }
         public DbSet<PostCategory> PostCategories { get; set; }
         public DbSet<Post> Posts { get; set; }