+++ src/Ops.DataProvider.SqlServer.Ops/Migrations/20201002173708_ops_v1.0.0.73.cs
@@ -1,0 +1,34 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+namespace Ocuda.Ops.DataProvider.SqlServer.Ops.Migrations
+{
+    public partial class ops_v10073 : Migration
+    {
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.CreateTable(
+                name: "PermissionGroupApplication",
+                columns: table => new
+                {
+                    ApplicationPermission = table.Column<string>(maxLength: 255, nullable: false),
+                    PermissionGroupId = table.Column<int>(nullable: false)
+                },
+                constraints: table =>
+                {
+                    table.PrimaryKey("PK_PermissionGroupApplication", x => new { x.PermissionGroupId, x.ApplicationPermission });
+                    table.ForeignKey(
+                        name: "FK_PermissionGroupApplication_PermissionGroups_PermissionGroupId",
+                        column: x => x.PermissionGroupId,
+                        principalTable: "PermissionGroups",
+                        principalColumn: "Id",
+                        onDelete: ReferentialAction.Restrict);
+                });
+        }
+
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropTable(
+                name: "PermissionGroupApplication");
+        }
+    }
+}